+++ docs/conf.py
@@ -1,6 +1,7 @@
 """
 Docs configuration
 """
+
 # -*- coding: utf-8 -*-
 #
 # capnp documentation build configuration file, created by
@@ -307,6 +308,6 @@ epub_copyright = "2013, Author"
 # Allow duplicate toc entries.
 # epub_tocdup = True
 
-intersphinx_mapping = {"http://docs.python.org/": None}
+intersphinx_mapping = {"<name>": ("http://docs.python.org/", None)}
 
 smv_branch_whitelist = r"^master$"